feat: fix wallet refresh issue
